Ignore ancillary files when copying to Docker image
This commit updates the .dockerignore file with some files that are not
relevant or useful in the context of a Docker image. This is a somewhat
trivial change, but prevents changes to these files from triggering
pointless layer rebuilds, and keeps the image's working directory more
tidy.

.github/: This contains a GitHub issue template, which is obviously not
used inside the container.

.gitignore: Since we are already ignoring the .git directory, this file
will have no effect.

.travis.yml: Similarly to .github/, this file is not used inside the container.

config.json.demo: We are already ignoring account-data/, so this demo
config will not be usable to my understanding.