Adjust Docker build to improve caching and reduce size
This commit makes some adjustments to the Dockerfile to avoid build
cache invalidation where possible, as well as some other more subjective
changes that I'll justify but would happily revert.

Firstly, I combined the apt-get commands into a single layer which also
deletes the package cache files, as shown in the Docker's best practices
documentation [1]. Per the same documentation, I separated the list of
packages into their own lines, sorted alphabetically, which makes updating
the list of packages easier in the future, and makes changes easier to diff.

I added an initial COPY of only the requirements.txt file before running
`pip install`, followed by a full COPY of all the source files. This means that
the requirements will only need to be downloaded, built, and installed
when the requirements change, rather than invalidating the cached layer
whenever any file is modified.

I removed the AWS_DEFAULT_REGION environment variable, as the user should
most likely be setting this explicitly anyways.

Finally, I removed some of the packages being installed that appeared to not be
useful in a Docker container:

automake: Installed automatically as a recommended dependency of autoconf.

python3.7-dev: The requirements appear to install successfully without
this, and since the python Docker image is built from source, I'm not
sure if this would even link properly.

jq: While jq is useful for parsing the output, users can install it themselves
if they need it (and can't just pipe the container output to jq on their local
machine).

awscli: Similarly to jq, this doesn't appear to be strictly necessary
for using cloudmapper.

[1] https://docs.docker.com/develop/develop-images/dockerfile_best-practices/#sort-multi-line-arguments